Split up p2panda-api
