+++ src/p2panda-api/entry.ts
@@ -1,0 +1,9 @@
+export type Entry = {
+  author: string;
+  encoded: string;
+  messageEncoded: string;
+  hash: string;
+  logId: number;
+  seqNum: number;
+  decoded: any;
+};
